+++ src/lib/agent/ChatAgent.ts
@@ -1,7 +1,7 @@
 import { ExecutionContext } from '@/lib/runtime/ExecutionContext'
 import { MessageManager } from '@/lib/runtime/MessageManager'
 import { ToolManager } from '@/lib/tools/ToolManager'
-import { createScrollTool, createScreenshotTool } from '@/lib/tools'
+import { ScrollTool, ScreenshotTool } from '@/lib/tools'
 import { generateSystemPrompt, generatePageContextMessage, generateTaskPrompt } from './ChatAgent.prompt'
 import { AIMessage, AIMessageChunk } from '@langchain/core/messages'
 import { PubSub } from '@/lib/pubsub'
@@ -56,8 +56,8 @@ export class ChatAgent {
    */
   private _registerTools(): void {
     // Only register the 2 essential tools for Q&A
-    this.toolManager.register(createScreenshotTool(this.executionContext))
-    this.toolManager.register(createScrollTool(this.executionContext))
+    this.toolManager.register(ScreenshotTool(this.executionContext))
+    this.toolManager.register(ScrollTool(this.executionContext))
     
     Logging.log('ChatAgent', `Registered ${this.toolManager.getAll().length} tools for Q&A mode`)
   }
