+++ src/lib/tools/GetSelectedTabsTool.ts
@@ -18,7 +18,7 @@ export const TabInfoSchema = z.object({
 
 export type TabInfo = z.infer<typeof TabInfoSchema>
 
-export class GetSelectedTabsTool {
+export class GetSelectedTabsToolImpl {
   constructor(private executionContext: ExecutionContext) {}
 
   async execute(_input: GetSelectedTabsInput): Promise<ToolOutput> {
@@ -58,15 +58,15 @@ export class GetSelectedTabsTool {
 }
 
 // LangChain wrapper factory function
-export function createGetSelectedTabsTool(executionContext: ExecutionContext): DynamicStructuredTool {
-  const getSelectedTabsTool = new GetSelectedTabsTool(executionContext)
-  
+export function GetSelectedTabsTool(executionContext: ExecutionContext): DynamicStructuredTool {
+  const tool = new GetSelectedTabsToolImpl(executionContext)
+
   return new DynamicStructuredTool({
     name: "get_selected_tabs_tool",
     description: "Get information about currently selected tabs. Returns an array of tab objects with id, url, and title. If no tabs are selected, returns the current tab.",
     schema: GetSelectedTabsInputSchema,
     func: async (args): Promise<string> => {
-      const result = await getSelectedTabsTool.execute(args)
+      const result = await tool.execute(args)
       return JSON.stringify(result)
     }
   })
