+++ src/lib/tools/Key.ts
@@ -3,45 +3,57 @@ import { z } from "zod";
 import { ExecutionContext } from "@/lib/runtime/ExecutionContext";
 import { PubSubChannel } from "@/lib/pubsub/PubSubChannel";
 
-const NavigateInputSchema = z.object({
-  url: z
-    .string()
-    .url()
-    .describe("Full URL to navigate to (must include https://)"),
+const KeyInputSchema = z.object({
+  key: z
+    .enum([
+      "Enter",
+      "Tab",
+      "Escape",
+      "Backspace",
+      "Delete",
+      "ArrowUp",
+      "ArrowDown",
+      "ArrowLeft",
+      "ArrowRight",
+      "Home",
+      "End",
+      "PageUp",
+      "PageDown",
+    ])
+    .describe("Keyboard key to press"),
 });
-type NavigateInput = z.infer<typeof NavigateInputSchema>;
+type KeyInput = z.infer<typeof KeyInputSchema>;
 
-export function createNavigateTool(
+export function KeyTool(
   context: ExecutionContext,
 ): DynamicStructuredTool {
   return new DynamicStructuredTool({
-    name: "navigate",
-    description: "Navigate to a URL",
-    schema: NavigateInputSchema,
-    func: async (args: NavigateInput) => {
+    name: "key",
+    description: "Send a keyboard key press",
+    schema: KeyInputSchema,
+    func: async (args: KeyInput) => {
       try {
         context.incrementMetric("toolCalls");
 
         // Emit thinking message
         context.getPubSub().publishMessage(
-          PubSubChannel.createMessage("Navigating...", "thinking")
+          PubSubChannel.createMessage("Pressing key...", "thinking")
         );
 
         // Get current page from browserContext
         const page = await context.browserContext.getCurrentPage();
 
-        await page.navigateTo(args.url);
-        await page.waitForStability();
+        await page.sendKeys(args.key);
 
         return JSON.stringify({
           ok: true,
-          output: `Successfully navigated to ${args.url}`,
+          output: `Pressed ${args.key} key`,
         });
       } catch (error) {
         context.incrementMetric("errors");
         return JSON.stringify({
           ok: false,
-          error: `Navigation failed: ${error instanceof Error ? error.message : String(error)}`,
+          error: `Key press failed: ${error instanceof Error ? error.message : String(error)}`,
         });
       }
     },
