+++ src/lib/tools/Type.ts
@@ -1,26 +1,33 @@
 import { DynamicStructuredTool } from "@langchain/core/tools";
 import { z } from "zod";
 import { ExecutionContext } from "@/lib/runtime/ExecutionContext";
+import { PubSubChannel } from "@/lib/pubsub/PubSubChannel";
 
-const ClickInputSchema = z.object({
+const TypeInputSchema = z.object({
   nodeId: z
     .number()
     .int()
     .positive()
     .describe("The nodeId number from [brackets] in element list"),
+  text: z.string().describe("Text to type into the element"),
 });
-type ClickInput = z.infer<typeof ClickInputSchema>;
+type TypeInput = z.infer<typeof TypeInputSchema>;
 
-export function createClickTool(
+export function TypeTool(
   context: ExecutionContext,
 ): DynamicStructuredTool {
   return new DynamicStructuredTool({
-    name: "click",
-    description: "Click an element by its nodeId (number in brackets)",
-    schema: ClickInputSchema,
-    func: async (args: ClickInput) => {
+    name: "type",
+    description: "Type text into an input element",
+    schema: TypeInputSchema,
+    func: async (args: TypeInput) => {
       try {
         context.incrementMetric("toolCalls");
+
+        // Emit thinking message
+        context.getPubSub().publishMessage(
+          PubSubChannel.createMessage(`Typing "${args.text}"...`, "thinking")
+        );
 
         // Get current page from browserContext
         const page = await context.browserContext.getCurrentPage();
@@ -36,18 +43,18 @@ export function createClickTool(
           });
         }
 
-        await page.clickElement(args.nodeId);
+        await page.inputText(args.nodeId, args.text);
         await page.waitForStability();
 
         return JSON.stringify({
           ok: true,
-          output: `Successfully clicked element ${args.nodeId} ${scrollMessage}`,
+          output: `Successfully typed "${args.text}" into element ${args.nodeId} ${scrollMessage}`,
         });
       } catch (error) {
         context.incrementMetric("errors");
         return JSON.stringify({
           ok: false,
-          error: `Failed to click : ${error instanceof Error ? error.message : String(error)}`,
+          error: `Failed to type into : ${error instanceof Error ? error.message : String(error)}`,
         });
       }
     },
