+++ src/lib/tools/index.ts
@@ -5,30 +5,30 @@ export * from './tool.interface'
 export * from './ToolManager'
 
 // Individual tools
-export * from './click'
-export * from './type'
-export * from './clear'
-export * from './scroll'
-export * from './navigate'
-export * from './key'
-export * from './wait'
-export * from './todo-set'
-export * from './todo-get'
-export * from './tabs'
-export * from './tab-open'
-export * from './tab-focus'
-export * from './tab-close'
-export * from './extract'
-export * from './human-input'
-export * from './done'
-export * from './celebration'
-export * from './visual-click'
-export * from './visual-type'
-export * from './click-at-coordinates'
-export * from './type-at-coordinates'
-export * from './grep-elements'
-export * from './screenshot'
-export * from './planner'
+export * from './Click'
+export * from './Type'
+export * from './Clear'
+export * from './Scroll'
+export * from './Navigate'
+export * from './Key'
+export * from './Wait'
+export * from './TodoSet'
+export * from './TodoGet'
+export * from './Tabs'
+export * from './TabOpen'
+export * from './TabFocus'
+export * from './TabClose'
+export * from './Extract'
+export * from './HumanInput'
+export * from './Done'
+export * from './Celebration'
+export * from './VisualClick'
+export * from './VisualType'
+export * from './ClickAtCoordinates'
+export * from './TypeAtCoordinates'
+export * from './GrepElements'
+export * from './Screenshot'
+export * from './Planner'
 export * from './MCPTool'
 export * from './GetSelectedTabsTool'
 export * from './GroupTabsTool'
